client

+++ rocket/net/tcp/tcp_buffer.cc
@@ -100,7 +100,7 @@ namespace rocket
 
     void TcpBuffer::moveReadIndex(int size)
     {
-        int j = m_read_index + size;
+        size_t j = m_read_index + size;
         if (j >= m_buffer.size())
         {
             ERRORLOG("moveReadIndex error,invalid size %d, old_read)index %d,buffer size %d", size, m_read_index, m_buffer.size());
@@ -112,7 +112,7 @@ namespace rocket
 
     void TcpBuffer::moveWriteIndex(int size)
     {
-        int j = m_write_index + size;
+        size_t j = m_write_index + size;
         if (j >= m_buffer.size())
         {
             ERRORLOG("moveWriteIndex error,invalid size %d, old_read)index %d,buffer size %d", size, m_read_index, m_buffer.size());
